+++ testData/org/elixir_lang/model/psi/variable/usages_variable_shadowed_in_case_clause.ex
@@ -1,0 +1,11 @@
+defmodule Shadowing do
+  def run(list) do
+    no<caret>de = 1
+
+    case list do
+      node -> node + 1
+    end
+
+    node
+  end
+end