+++ testData/org/elixir_lang/model/psi/variable/usages_variable_shadowed_in_comprehension_do_block.ex
@@ -1,0 +1,11 @@
+defmodule Shadowing do
+  def run(list) do
+    no<caret>de = 1
+
+    for node <- list do
+      node + 1
+    end
+
+    node
+  end
+end