+++ testData/org/elixir_lang/model/psi/variable/usages_variable_shadowing_anonymous_function_parameter.ex
@@ -1,0 +1,7 @@
+defmodule Shadowing do
+  def run(list) do
+    node = 1
+    Enum.map(list, fn no<caret>de -> node + 1 end)
+    node
+  end
+end